# Conflicts:
#	docs/en/interfaces/formats.md
#	docs/en/operations/settings/settings-formats.md
#	src/Core/Settings.h
#	src/Core/SettingsChangesHistory.h
#	src/Formats/FormatFactory.cpp
#	src/Processors/Formats/Impl/CSVRowInputFormat.cpp

+++ src/Interpreters/AsynchronousMetricLog.cpp
@@ -1,4 +1,3 @@
-#include <base/getFQDNOrHostName.h>
 #include <DataTypes/DataTypeDate.h>
 #include <DataTypes/DataTypeDateTime.h>
 #include <DataTypes/DataTypeDateTime64.h>
@@ -6,8 +5,10 @@
 #include <DataTypes/DataTypeString.h>
 #include <DataTypes/DataTypesNumber.h>
 #include <Interpreters/AsynchronousMetricLog.h>
-#include <Parsers/parseQuery.h>
 #include <Parsers/ExpressionElementParsers.h>
+#include <Parsers/parseQuery.h>
+#include <base/getFQDNOrHostName.h>
+#include <Common/DateLUTImpl.h>
 
 
 namespace DB
@@ -67,7 +68,7 @@ void AsynchronousMetricLog::addValues(const AsynchronousMetricValues & values)
     AsynchronousMetricLogElement element;
 
     element.event_time = time(nullptr);
-    element.event_date = DateLUT::instance().toDayNum(element.event_time);
+    element.event_date = static_cast<UInt16>(DateLUT::instance().toDayNum(element.event_time));
 
     /// We will round the values to make them compress better in the table.
     /// Note: as an alternative we can also use fixed point Decimal data type,
